cloudapi: split up service into logical sections
